Move keymap.c to max/keymap.c , Create pristine default keymap.c,
